rename scenario
