+++ tests/end-to-end/external-solution-integration/internal/scenario/send_and_check_event/scenario.go
@@ -1,4 +1,4 @@
-package send_and_check_event_only
+package send_and_check_event
 
 import (
 	"github.com/kyma-project/kyma/tests/end-to-end/external-solution-integration/internal/scenario"
